refactor event logic into module and add tests
