+++ web_gui/App.jsx
@@ -1,5 +1,6 @@
 import { useEffect, useRef, useState } from 'react';
 import GameBoard from './GameBoard.jsx';
+import { applyEvent } from './applyEvent.js';
 import './style.css';
 
 export default function App() {
@@ -56,35 +57,6 @@ export default function App() {
     }
   }
 
-  function applyEvent(state, event) {
-    if (!state) return state;
-    const newState = JSON.parse(JSON.stringify(state));
-    switch (event.name) {
-      case 'start_game':
-        return event.payload.state;
-      case 'draw_tile': {
-        const p = newState.players[event.payload.player_index];
-        if (p) p.hand.tiles.push(event.payload.tile);
-        if (newState.wall?.tiles?.length) newState.wall.tiles.pop();
-        break;
-      }
-      case 'discard': {
-        const p = newState.players[event.payload.player_index];
-        if (p) {
-          const { tile } = event.payload;
-          const idx = p.hand.tiles.findIndex(
-            (t) => t.suit === tile.suit && t.value === tile.value,
-          );
-          if (idx !== -1) p.hand.tiles.splice(idx, 1);
-          p.river.push(tile);
-        }
-        break;
-      }
-      default:
-        break;
-    }
-    return newState;
-  }
 
   function handleMessage(e) {
     try {
